The program works and creates the sums per the instructions
The program works and creates the sums per the instructions. The missing details are the inclusion of the Assert Statements with the appropriate output. This is the 2nd version of the program and uses a list to build the sums.
The type of sum series to be built is dependent on the inputs from the key board. If the number terms in the series is the only entry (use defaults) for the 2nd two inputs, a Fibonacci series is generated. If the 1st two elements are input as "2" and "1", a Lucas series is created. Otherwise a general sum series is created.
20220211