The file series.py cpntains an alternate method of generating sum series
The method used in series.py builds the output from small number to larger ones using recursion from the definition instead of the math formula Fn = (n-1) + (n-2). It was easier to visualize what was done and the output is in a list with all of the generated values. This program was not turned as the assignment.
20220210